list is mutable

+++ 0x02-csharp-arrays_lists_dictionaries/7-delete_at/7-delete_at.cs
@@ -4,7 +4,11 @@ class List
 {
     public static System.Collections.Generic.List<int> DeleteAt(System.Collections.Generic.List<int> myList, int index)
     {
-        myList.RemoveRange(index, 1);
+        if (index >= 0 && index < myList.Count)
+        {
+            myList.RemoveRange(index, 1);
+        }
+
         return myList;
     }
 }
